Fix unreliable tests (#1287)
Summary:

There were a few CI test failures due to recently landed diffs:
- https://app.circleci.com/pipelines/github/facebook/hermes/5538/workflows/c993258e-511e-4622-b932-973d41f75fd1/jobs/55809
- https://app.circleci.com/pipelines/github/facebook/hermes/5536/workflows/09c1b07c-b12e-4cc7-9a66-97ee30bc7058/jobs/55789

Both failures are due to:
```
C++ exception with description "triggerInterrupt didn't get executed" thrown in the test body.
```

I'm not able to reproduce the test failure locally, but suspect that 
this is due to the `waitFor` of explicit pause timinig out because 
there isn't JavaScript to run. If true, the failure is intermittent 
based on timing of the VM executing the pre-fix JavaScript code.

Reviewed By: mattbfb

Differential Revision: D53204781