Allow pre-enabling the Runtime domain in CDPHandler
Summary:
When integrating Hermes's `CDPHandler` into React Native, the Hermes runtime doesn't necessarily exist for the entire CDP session, and in particular can be destroyed and recreated due to a reload. For similar reasons, `CDPHandler` can be destroyed and recreated.

If `CDPHandler` is recreated after a `Runtime.enable` message has already been received, the client will be expecting a `Runtime.executionContextCreated` notification, but `CDPHandler` will not send it unless the client sends an additional `Runtime.enable`. Based on Chrome's behaviour, this is incorrect: `Runtime.enable` should effectively be Session-scoped rather than Agent-scoped.

Therefore, we need a way to pre-enable the Runtime domain when instantiating `CDPHandler`. This diff proposes one possible API: passing an (optional) session config object to the constructor. We also need to send the notification as a side effect of `registerCallbacks` in the case where the Runtime domain is pre-enabled.

Reviewed By: dannysu

Differential Revision: D53006914

fbshipit-source-id: e89405d9cdb380188649e6847d59bef40cb7bb6b