Fix GrammarContext for 'as' expression type
Summary:
The RHS of the `as` should be parsed with `Type` GrammarContext,
which can be done by moving the `advance()` call after the check.

Tested on a large Fb4a bundle and this doesn't appear to regress overall
parser performance at all (no extra work is done, just some reordering).

Reviewed By: tmikov

Differential Revision: D53135103

fbshipit-source-id: f24ad7482aa743ff5826460cded4df81d97e7db7