Fixing calculation of frame size when debugging (#1275)
Summary:
I'm encountering a crash when the Hermes Engine hits a break point in certain code configurations, and tries to return the Call Frames through the CDPHandler. This is reproducible 100% of the time for me.

When a breakpoint is hit, `Debugger::getLexicalInfoInFrame` is called and the following code runs:
```
  uint32_t currFrame = 0;
  while (auto idx = getScopeDescIndexForFrame(scopeDescs, currFrame++)) {
    result.variableCountsByScope_.push_back(getFrameSize(scopeDescs, *idx));
  }
```

`getScopeDescIndexForFrame`, will just return the inner-most scope index for the frame, so that's the 0th scope for frame 0.

However when `getFrameSize` is called in certain cases, it will crash.
For example with the following code:
```
function testFunction(num: number) {
  if (num > 1) {
    try {
      console.log('HelloWorld2' + num);
    } catch (err: any) {
      console.log('HelloError');
    }
  }
}
```
The scopeDesc flags will look like:
`scopeDesc[0].isInnerScope = true` (The try/catch scope)
`scopeDesc[1].isInnerScope = true` (The if scope)
`scopeDesc[2].isInnerScope = false` (The function scope)

`getScopeDescIndexForFrame(scopeDescs, 0)` will return *idx=0
This then gets passed to `getFrameSize(scopeDescs, 0)` which contains the following code:
```
  do {
    frameSize += scopeDescs[i].names.size();
  } while (scopeDescs[i--].flags.isInnerScope);
```

Which will crash on the 2nd iteration since `scopeDescs[0].flags.isInnerScope == true` so it will then try to access `scopeDescs[-1]`. I believe this should be `i++`, to count all of the sizes of scopes from inner-most to outer-most, stopping after a non-innerScope, indicating that the next scope is a new frame. In this case, it should be adding up the `names.size()` of scopes, 0, 1 and 2, instead of starting at 0 and counting down to -1.

Pull Request resolved: https://github.com/facebook/hermes/pull/1275

Test Plan: Just CI Tests.

Reviewed By: dannysu

Differential Revision: D53137420

Pulled By: mattbfb

fbshipit-source-id: 70f8c3474fc9cf32e1a925bc73292e44c1c530ac