Add Unicode properties support to genUnicodeTable.py
Download a significant portion of the Unicode database in order to
generate codepoint tables, and property name/value canonicalization,
for the Unicode 15.1.0 binary and non-binary property names/aliases
supported by ES262.

Certain properties that are not explicitly defined in the Unicode
database are derived here, such as "Assigned", and "Any".