Add support for Unicode property names to the RegExp parser
In Unicode mode, RegExp is now able to parse `\p{...}` and `\P{...}`
atom escapes, both in and out of character classes.

Outside of Unicode mode, `\p` and `\P` are unnecessary escapings of
literal `p` and `P` respectively.

A new error type `InvalidPropertyName` was added, which is returned
for all property-name related parsing/resolving issues. This seems to
be consistent with other JS engines.