Print labels and module in IRVerifier
Summary:
Upon the first failure, IRVerifier will dump the entire module.
AssertIWithMsg will then be able to reference the labels used in the
dump to generate error messages that pinpoint the exact location of the
failing instruction. Because IRVerifier is now responsible for dumping
the module, remove the call to dump the module in SH.cpp.

Reviewed By: avp

Differential Revision: D53363244

fbshipit-source-id: 75666571fed9e374d28447f842f406299b87c50f