Handle eval command & EvalComplete differently
Summary:
It does not make sense to broadcast the `EvalComplete` event to all
subscribers because the EvalComplete event is for a specific request.

Performing an eval as the next command from didPause is actually a bit
special. It's different because `lib/VM/Debugger/Debugger.cpp` will
remain in the debugger loop and any expression that gets evaluated will
not trigger recursive `didPause`. So the result of running an eval
command and getting the EvalComplete back can be seen as if those 2
things are just a synchronous function and the PauseReason has not
changed.

Reviewed By: mattbfb

Differential Revision: D53257099

fbshipit-source-id: f76cd63b8adea459f25b776872e7b2ab96d04102