Add ESTreeClone
Summary:
Add an AST cloning function that also copies over all relevant
SemContext structures.

This requires adding some internal cloning functionality in SemContext
and in `DeclCollector` (because FlowChecker wants that as well).

The cloning function is just an AST visitor that clones recursively,
with functions before and after the clone to setup/finish cloning the
semantic data.
There's a minor amount of type shenanigans required for `NodeList`,
because it's not copyable.
The cloner keeps a couple of maps to map from old->new data as it
clones.

This function will be used for cloning ASTs to specialize generics
when FlowChecker encounters uses of generics.

Reviewed By: tmikov

Differential Revision: D49884370

fbshipit-source-id: 4dfe7a8d06235ad1c0fd845fe317836bdb3997e6