Add BaseScopeInst
Summary:
Add a new base class `BaseScopeInst` to represent operations that
produce a scope. This is primarily a convenience, to make it easy to
check the incoming scope to an operation.

Some examples of things that this is used for

1. Verify that the incoming `VariableScope` to a load/store is the
right one for the `Variable` being accessed, and that the
`VariableScope` provided during closure creation is actually the parent
of the child function.
2. Optimise out a `ResolveScopeInst` if its incoming scope is the same
as the one we want.

As a start, use for the existing HBC instructions.

Reviewed By: avp

Differential Revision: D53394900

fbshipit-source-id: 4d826b8acfcb7bef138a46b065ce55a403938a2a