Use explicit environments in emitted code
Summary:
Add more explicit operations that manipulate scopes in SH generated
code. This allows us to retrieve the parent scope inline in native
code, and then call into resolve/create scope operations that take that
scope as an operand.

This is necessary to support additional scopes in functions, since
their parents are not necessarily the parent of the current function.

Reviewed By: avp

Differential Revision: D53507852

fbshipit-source-id: 2f6fff0b3b937b96514005ac093ff5cecd8ab106