Implement Debugger.setBreakpointByUrl
Summary:
Port over the `Debugger.setBreakpointByUrl` implementation from
CDPHandler.

The code in CDPHandler breaks the processing of new script and sending
`breakpointResolved` notification into two steps:
`processCurrentScriptLoaded()` and `processPendingScriptLoads()`.

In the new code, this is handled in a more intuitive way and completely
contained in `processNewLoadedScript()`. The new code no longer
maintains a list of loaded scripts because it can simply obtain it from
DebuggerAPI via the `getLoadedScripts()` added by D51684983. That's why
the new code doesn't have the equivalent of
`processPendingScriptLoads()`.

Reviewed By: mattbfb

Differential Revision: D53550169

fbshipit-source-id: 08e554b2d0d4938db12419da9d29d850db1ce290