Allow generic instantiations in type aliases
Summary:
Allowing generics on the right hand side of type aliases requires
deferring the resolution of the generic classes until after type aliases
have all had types assigned to them.
Assign a Generic placeholder TypeInfo for these classes,
and use a side table of args needed to resolve them to a real ClassType.

The generic classes are then completed at the same time as the unions,
and their TypeInfo is replaced by the actual ClassType.

Differential Revision: D51644093

fbshipit-source-id: 5bcb70b83a7b75d72f101384cd3d2921c328fd2f