Allow generic instantiations in type aliases
Summary:
Allowing generics on the right hand side of type aliases requires
deferring the resolution of the generic classes until after type aliases
have all had types assigned to them.
Assign a Generic placeholder TypeInfo for these classes,
and use a side table of args needed to resolve them to a real ClassType.

The generic classes are then completed at the same time as the unions,
and their TypeInfo is replaced by the actual ClassType.

Reviewed By: tmikov

Differential Revision: D51644093

fbshipit-source-id: cc6d89c7d1912cddb207a37d17ed70acc53555ef