Removing API usage not applicable on iOS (stat and fstat) in libhermes
Summary:
This diff just removes references to functions that use (either
directly or transitively) those functions, and relies on linker DCE
to clean up the rest. Trying to extricate all uses of these functions
would otherwise result in very significant changes.

This diff makes two changes:
1. In raw_fd_ostream, fall back to a default buffer size when building
for iOS.
2. Update a function in LLVM that was marked as "used" to suppress
warnings, to instead use the new `[[maybe_unused]]` attribute. This
allows the function, and the calls to stat it retains, to be stripped.

Reviewed By: tmikov, TheSavior

Differential Revision: D53133166

fbshipit-source-id: f425762bc6b7072cbd8235c65e9cf652cecb981c