Allow stripping of file cleanup code
Summary:
The linker currently cannot strip the file cleanup code because it is
called on fatal errors and signals (when a the LLVM signal handler is
installed). To allow it to be stripped, add a layer of indirection
through a function pointer, which is only populated by
`RemoveFileOnSignal`. This way, the cleanup code is only considered
reachable when `RemoveFileOnSignal` is actually used, and it can be
stripped otherwise.

Reviewed By: tmikov

Differential Revision: D53535800

fbshipit-source-id: ae26197a66bc154dd83eb304b4c731eabac10873