Fix Windows builds in CircleCI
Summary:
For locating cmake:
`refreshenv` for chocolatey requires extra setup for powershell:
https://docs.chocolatey.org/en-us/troubleshooting#refreshenv-has-no-effect

Since we know where it gets installed to, we can just update the env
ourselves.

For python3:
It looks like somehow the machines have 3.11 installed instead of 3.12.
Don't know how that happened. In any case, unifying the setup between
the "windows" and "test-windows" jobs is probably a good idea. And
doing a choco install of python3 seems allow test-windows to pass.

Reviewed By: neildhar

Differential Revision: D53682818

fbshipit-source-id: 0eb267ffc18219895de4b7f7edf4d0b5a9fe8ecf