Added support for providing bytecode generation option to BytecodeProviderFromSrc (#1262) (#1262)
Summary:
Original Author: simon@corsin.me
Original Git: 1547bd461ed8075552f59f76b23de9456af0bbd4
Original Reviewed By: tmikov
Original Revision: D52918720

This change adds support for providing the bytecode generation options
to the `BytecodeProviderFromSrc` API. This makes it possible to strip
the function names and debug information when using the high level
`BytecodeProviderFromSrc` API without having to re-write what
`BytecodeProviderFromSrc` or `CompilerDriver` are already doing. For
context, we currently statically embed Hermes into one of our own CLI
binary that is able to do a lot of different compile related tasks for
us, as such we are not using the Hermes CLI API. We still would like to
be able to compile modules as byte code, and the only thing missing
from `BytecodeProviderFromSrc` to be at parity with the
`CompilerDriver` CLI option that does bytecode generation was the
ability to configure the bytecode generation phase.

Pull Request resolved: https://github.com/facebook/hermes/pull/1262

Pulled By: neildhar

Reviewed By: avp

Differential Revision: D54092028

fbshipit-source-id: efbeb8a1d02e0cd83f239515f7101012d6414b81