Replicate inspector/chrome files to cdp
Summary:
My previous plan (D53839508) was to expose the common classes from the
`HermesAPI` target, then have the `API/hermes/inspector:chrome` target
consume those from `HermesAPI`. This way I can have the new CDPAgent
and the old inspector/chrome classes share code. With the ultimate plan
to still have the new CDPAgent be consumed from `HermesAPI`.

However, I hit a major blocker with that approach. It turns out that
several of the codegen classes in `MessageTypes.h` hit compilation
problems with MSVC if exported. For example, `debugger::Location`. This
class is marked to have copy constructor and copy assignment operator
deleted. However, it is being used inside a std::vector in
`SetBreakpointByUrlResponse` class. If `SetBreakpointByUrlResponse` is
marked as exported, then MSVC compilation fails due to
`debugger::Location` not having copy constructor and copy assignment
operator.

So I'm taking a different approach by replicating stuff in
`inspector/chrome` to the `cdp` directory to be used by the new
`CDPAgent` and related classes. I used `hg cp` so the history should be
linked. The pro of this approach is that it keeps the exported symbols
of `HermesAPI` clean. The down side is that until we get rid of
`CDPHandler` and the associated `API/hermes/inspector:chrome` target,
there is some code duplication.

That said, we have a defined time schedule for when we're planning to
get RN off of `CDPHandler`. This situation should not persist past Q3.
I think this still gets the BUCK build to the same shape as our cmake
build and ultimately better.

Reviewed By: mattbfb

Differential Revision: D53966898

fbshipit-source-id: 720d2e05c670e7e820a9488d7e1e80acd8e4a760