Allow stripping of file cleanup code
Summary:
Original Author: neildhar@meta.com
Original Git: b686d8f4da2f0341346a761fadabff8253a7059f
Original Reviewed By: tmikov
Original Revision: D53535800

The linker currently cannot strip the file cleanup code because it is
called on fatal errors and signals (when a the LLVM signal handler is
installed). To allow it to be stripped, add a layer of indirection
through a function pointer, which is only populated by
`RemoveFileOnSignal`. This way, the cleanup code is only considered
reachable when `RemoveFileOnSignal` is actually used, and it can be
stripped otherwise.

Reviewed By: avp

Differential Revision: D54092022

fbshipit-source-id: 050b7384c5453ee735b33411ef9f828dbfdfc1d6