Mirror HBC scope instructions in non-lowered IR
Summary:
This diff eliminates the HBC scope instructions in favour of ordinary
IR instructions that are introduced directly by IRGen.

Concretely:

1. `HBCLoadFromEnvironmentInst` is replaced by adding a scope parameter
to `LoadFrameInst`
2. `HBCStoreToEnvironmentInst` is replaced by adding a scope parameter
to `LoadFrameInst`
3. `HBCResolveEnvironmentInst` is replaced by a new instruction
`ResolveScopeInst`, which is emitted by IRGen
4. `HBCCreateFunctionInst` is replaced by adding a scope parameter to
`CreateFunctionInst`
5. `HBCCreateEnvironmentInst` is replaced by a new instruction
`CreateScopeInst` which is currently created for each function in
`IRGen`

Note that this should not be a functional change, we are effectively
just eliminating the lowering pass `LowerLoadStoreFrameInst` and
directly emitting the lowered IR at IRGen.

Differential Revision: https://internalfb.com/D46531173