Fix arguments-decrement.js test and delete IR pass test
Summary:
This test wasn't testing the `updateIncomingPhiValues` call
that it was originally intended to test, because SH's modified compiler
pipeline was removing the Phi early.

Add a more explicit second Phi to more properly test this.

The test that checked IR after every pass also goes away,
because the error is now caught by IRVerifier so there's no reason to
have this huge codegen test that creates a lot of noise.

Reviewed By: neildhar

Differential Revision: D54376668

fbshipit-source-id: e52a3d1d6f97e6bc1d101091232064ee59c28f70