Add instruction to get env from closure
Summary:
Add an instruction that reads the environment from a given closure.
This is necessary for inlining, since a function may be inlined into a
function from which its variables are not visible.

For example:
```
function foo(){
var i = 0;
function bar(){
return ++i;
}
return bar;
}

var b = foo();
b();
```

Inlining the call to `b()` will require retrieving the scope in which
`i` lives from the closure.

Reviewed By: avp

Differential Revision: D53462041

fbshipit-source-id: a502eb64cef59f7e6882b4acedac19392570562e