Perform scope analysis directly on scopes
Summary:
Eliminate FunctionScopeAnalysis, and instead add a simple recursive
lookup in `LowerScopes` that computes scope depth purely by walking the
scope chain. This allows us to decouple the depth of a scope from the
depth of a function, which is necessary to support multiple scopes, and
to eliminate scopes that have no variables.

Reviewed By: avp

Differential Revision: D53670070

fbshipit-source-id: 7ecfacfc291a6a5e9bd3637bed3bdea5b23ff864