Mirror HBC scope instructions in non-lowered IR
Summary:
This diff eliminates the HBC scope instructions in favour of ordinary
IR instructions that are introduced directly by IRGen.

Concretely:

1. `HBCLoadFromEnvironmentInst` is replaced by adding a scope parameter
to `LoadFrameInst`
2. `HBCStoreToEnvironmentInst` is replaced by adding a scope parameter
to `StoreFrameInst`
3. `HBCResolveEnvironmentInst` is replaced by a `ResolveScopeInst`,
which is emitted by IRGen
4. `HBCCreateFunctionInst` is replaced by adding a scope parameter to
`CreateFunctionInst`
5. `HBCCreateEnvironmentInst` is replaced by a `CreateScopeInst`.

The new instructions differ slightly from their HBC counterparts in
that environment creation and resolution take explicit scope
parameters. This is necessary to easily model multiple scopes being
simultaneously live in a function.

Allow the changed instructions to pass through lowering, and add code
emission in the backend to handle them.

The now unused HBC instructions are deleted in subsequent diffs.

Reviewed By: avp

Differential Revision: D46531173

fbshipit-source-id: a8916f629e3a904b4821a4bb7d5646d3fcba2771