Allow using `babel-plugin-syntax-hermes-parser` with Babel 8 alpha (#1330)
Summary:
<!--
  Thanks for submitting a pull request!
  We appreciate you spending the time to work on these changes. Please provide enough information so that others can review your pull request. The two fields below are mandatory.

  Before submitting a pull request, please make sure the following is done:

  1. Fork [the repository](https://github.com/facebook/hermes) and create your branch from `main`.
  2. If you've fixed a bug or added code that should be tested, add tests!
  3. Ensure it builds and the test suite passes. [tips](https://github.com/facebook/hermes/blob/HEAD/doc/BuildingAndRunning.md)
  4. Format your code with `.../hermes/utils/format.sh`
  5. If you haven't already, complete the CLA.
-->

The plugin currently requires Babel 7. Babel 8 alpha has no changes to the API that the plugin is using, so it will work if we allow it.

Note that there are some AST breaking changes in Babel 8, so the plugin will have to be updated to fully support it: https://next.babeljs.io/docs/v8-migration-api#ast-changes. However, this PR at least allows experimenting with the plugin in Babel 8 (and making sure that everything else works) for people not affected by those AST changes.

Pull Request resolved: https://github.com/facebook/hermes/pull/1330

Test Plan:
<!--
  Demonstrate the code is solid.
  Example: The exact commands you ran and their output,
  screenshots / videos if the pull request changes the user interface.
-->

Existing tests pass.

Reviewed By: SamChou19815

Differential Revision: D54316701

Pulled By: pieterv

fbshipit-source-id: a1bf9fb66ebb12fc82294a19de73edefff0a5e31