Allow Debugger to compile by stubbing out some functions
Summary:
Compiling with `HERMES_ENABLE_DEBUGGER` didn't work because we haven't
been enabling it for a while, so try enabling it and putting in partial
implementations of the parts that have changed.

This doesn't attempt to actually fully support the Debugger,
given that that would require eval, ScriptID support for SHUnit, etc.
It just makes sure the file can compile as a first step.

Disable the hdb debugger tests for now because many of them fail
due to the `eval` problems.

Reviewed By: neildhar

Differential Revision: D54452213

fbshipit-source-id: d7810440472277142dfcdfdbdc4a16e19ca25636