Hook up BUCK test for hermes_parser
Summary:
The main goal is to setup a BUCK workflow to speed up iterations. This would also enable us to set up CI on sandcastle.

Currently there is an issue with the insta crate we are using to generate snapshots, it would treat the same fixture as different under cargo and BUCK because the base dir is different. To get around it, we create a different snapshot folder for internal.

Reviewed By: captbaritone

Differential Revision: D54393704

fbshipit-source-id: 34689ad23760e08491805ef9349d61086a83bada