Change CDPAgent::getState to return unique_ptr
Summary:
For the integration in React Native (D54369985), we are using `CDPAgent` in a similar way to how we currently using `CDPHandler`.

Our `RuntimeAgentDelegate` interface expects a `std::unique_ptr`, as the expectation is that we take the final state of `CDPHandler`/`CDPAgent` at destruction, then pass ownership into the recreated instance after reload.

https://www.internalfb.com/code/fbsource/[85f1004d92b1]/xplat/js/react-native-github/packages/react-native/ReactCommon/jsinspector-modern/RuntimeAgentDelegate.h?lines=46

In this diff, I'm aligning the return type of `CDPAgentImpl::getState` to restore this expectation and fix compatibility.

Reviewed By: dannysu

Differential Revision: D54415953

fbshipit-source-id: 1c5af52a54555859cef5c23baf85ce5c247335e8