Prevent regex native stack overflow (#1191)
Summary:
Pull Request resolved: https://github.com/facebook/hermes/pull/1191

Currently, the regex executor has no bounds on the amount of recursive
calls it can make. This means that the regex executor can overflow.  This
diff refactors the logic that the Runtime class already has for guarding
against overflow into a separate class, and then uses that class in the
executor. This overflow checking class uses either real native stack checking or
a simple depth counter configured with a max depth allowed. The variant
is chosen based on `HERMES_CHECK_NATIVE_STACK`.

When stack checking is off, the depth counter is inherited from the
runtime. That's defined using a set of heuristics based on platform and build
mode. However, we set the max limit of the depth counter to a larger value,
since the stack size of the regex executor is less than the size of the
interpreter stack.

Reviewed By: dannysu

Differential Revision: D50425266

fbshipit-source-id: 6434e4320268e39a6857f73abc6f643dcbf12a8d