Re-enable ProfilerBasicOperation (#1326)
Summary:
Pull Request resolved: https://github.com/facebook/hermes/pull/1326

The old code had several issues.

First, the `registerForProfiling()` and `unregisterForProfiling()`
functions should be called on the runtime thread. CircleCI doesn't
support privileged mode for their docker environment. This means when
the test hangs, I'm unable to attach lldb to it to debug. But I was able
to reproduce the issue outside of CircleCI in a Linux VM with docker. It
looked like the `unregisterForProfiling()` call might have messed up
states that then caused the `SamplingProfiler::resume()` to not
complete. This means the runtime thread is blocked. SerialExecutor's
destructor will wait for the runtime thread to finish and so once the
runtime thread is blocked, the test will never finish due to being stuck
in TearDown.

Second, the code to set `stopFlag_` needs to be guaranteed to run in
all situations. Otherwise the same issue might occur where
SerialExecutor cannot finish its destructor.

Lastly, the test had checks that aren't deterministic. The sampling
profiler doesn't guarantee it will gather any particular samples within
the short run of the test, so certain checks need to be removed.

Reviewed By: mattbfb

Differential Revision: D54276786

fbshipit-source-id: 99b5644c4cddbc3bde8244a7d9e4abaa9b977b2f