Move creation of FunctionInfo for implicit constructor into SemanticResolver.
Summary:
In D54313169, I had been making a new FunctionInfo in
ESTreeIRGen-class.  Aakash requested that I do this instead in
SemanticResolver, for various architecture and hygiene reasons, and
save the FunctionInfo in the ClassLikeDecorator, to be used later in
ESTreeIRGen-class.  I committed to making that change in a later diff.
This is that diff.

To do this, I needed to know, in SemanticResolver, whether a
ClassLikeNode had a constructor.  So I factored out code in
ESTreeIRGen-class that searches for a constructor, and put it in
SemContext.  (I had originally thought this was a pure ESTree
operation, but then I found that it needed Keywords, which are
apparently constructed in SemContext.)

Reviewed By: avp

Differential Revision: D54595385

fbshipit-source-id: eae21ecf1b8d36693bbffc9f76f019b7c0fb9e61