Extend Semantic Resolver to check field initializers.
Summary:
This diff does semantic resolution on field initializers.  When a
class has such initializers, it creates a FunctionInfo for the
implicit field initializer function, with the right scope, and stores
that in a new decoration of class-like nodes.  This is used for all
the field initializers (and later will be used in IRGen).

This FunctionInfo is passed a null Function*; that argument was only
used to decide whether the FunctionInfo was for an arrow function.  We
ensure that null translates to false for this property.

We create ClassContext, to hold information about the current class in
SemanticResolver, and create one when we process class declaration and
expression nodes.

Reviewed By: avp

Differential Revision: D54443508

fbshipit-source-id: 3682dbd7a0c555dbc5d9ad9fab8ed33fa503e38f