Field initializers.
Summary:
Handle class field initializers properly.  When a class has fields
with initializers, we generate code for an implicit field initializer
function, and emit one or more calls to this in the constructor of the
class (including in the implicit constructor, for a class with no
explicit constructor).  If the class has no superclass, the call is at
the start of the constructor, before any explicit code.  If the class
does have a superclass, the call occurs immediately after any
"super()" call.  (These are currently required to be the first
statement of a constructor, but the code for emitting the field
initializer call should be robust in the face of future generalization
of this.)

As part of this, we add a ctorClass_ to FunctionContext -- the class
when the FunctionInfo is for an implicit function of the class.

ESTReeIRGen gets a classFieldInitInfo_ map, from ClassDeclarationNode*
to ClassFIeldInitInfo.  Currently, that consists just of the Variable*
into which we store the field initializer function.

ESTreeIRGen::ExtraKey gets a new member, ImplicitFieldInitializer.
That made the enum too large to fit in the alloted two bits, so we
separate the enum into values that apply to functions and to classs,
and number each set starting at 0.  That obeys the bit-width
restriction.

Reviewed By: avp

Differential Revision: D54263103

fbshipit-source-id: bf8fd1a46840244b6f4b63b5bac5a0de880196b2