In SemanticResolver, handle super calls that are in arrow functions nested in a ctor.
Summary:
Normally, super calls have to be in the derived class constructor.  But there is an exception: a constructor may define arrow functions, and super() within those is legal, and calls the super of the containing constructor.

This diff fixes the SemanticResolver to recognize these exceptions, and not give an error message.

(I don't know why error functions are different from nested ordinary functions in this regard, and can't find the precise rule in the spec.  But I can confirm that V8 rejects super() within a non-arrow nested function, so I assume this rules is correct.)

Reviewed By: avp

Differential Revision: D53978843

fbshipit-source-id: 044012d40b9dd2ffb60b024ca89e46bc1ee04de9