Allocate HashMapEntry on OldGen (#1344)
Summary:

Allocate HashMapEntry if the hash table is on OldGen. This gives us
much better performance when large number of entires (like 10M entries)
are inserted to Set or Map.
At early insertion, everything will be allocated on YoungGen, then at
some point, YoungGen collection happens and the hash table will be
moved to OldGen. After that, all the allocation and rehash will happen
on OldGen.

Reviewed By: neildhar

Differential Revision: D53293826