Message response should be at the end
Summary:
Per https://chromium.googlesource.com/chromium/src/+/master/third_party/blink/public/devtools_protocol/

> If a command invocation results in sending some events (for example,
the enable command of certain domains may result in sending of
previously buffered events), these are guaranteed to be emitted before
the method returns.

The response to a command should always be at the very end after sending
any notifications. The way V8 has implemented their interfaces
guarantees this. The response is always a return value of whatever
function that processes it. It is then up to a separate code to take the
response and craft the actual response payload.

For us, a better guarantee might be easier to tackle when we solve the
dispatch logic TODO. For now, just ensure we send the response at the
end.

Reviewed By: mattbfb

Differential Revision: D54758755

fbshipit-source-id: 447888d0040b02cbdec1f4cbab12fdb286c9e964