Suppress TSAN errors in JSI value reference counts
Summary:
TSAN complains when freeing JSI values from a background thread, because
we are using relaxed operations and depending on a load-store control
dependency to provide ordering. We previously had a TSAN suppression
here, but this was dropped when we switched over to ManagedChunkedList.

Instead of using a TSAN suppression again, which can be difficult to
place precisely (and may have a performance impact), just provide
ordering when building under TSAN. This makes the currently implicit
ordering explicit for TSAN.

Given the problems we have had with this, it may be worth quantifying
the actual performance benefit we get from using relaxed operations
instead of acquire/release in all cases.

Reviewed By: tmikov

Differential Revision: D54832907

fbshipit-source-id: 3b305e3a7f372c67b0872b05acfe9b8c0ff8a948