New class for storing domain state
Summary:
The current way of dealing with fetching domain state is by placing a
lock on all operations of `DomainAgents`. That potentially has issues
with re-entrancy if while processing a `handleCommand`, the code
triggers `DomainAgents` functions to be called again.

`recursive_mutex` might be one way to deal with it, but that's really
undesirable to have due to the permanent added complexity that
`recursive_mutex` brings. Also, clang's Thread-Safety Analysis can't be
applied to `recursive_mutex`.

This new `DomainState` class aims to narrow the scope of where locking
is done. Instead of locking operations that touch domain agents, the
`DomainState` class manages thread-safety within its operations such
that domain agents don't need to worry about it.

Reviewed By: mattbfb

Differential Revision: D54810302

fbshipit-source-id: 0c745a34a220e1c6e79b01c6ba6e4fdb48878957