Explicitly handle overflow in construction of function header
Summary:
Instead of implicitly overflowing in `SmallFuncHeader`'s constructor,
add separate constructors for the overflow and non-overflow case, and a
helper to determine if a given header will fit.

This gives the caller control over whether to emit a full size function
header, even if there isn't an overflow. This will be used in
subsequent diffs to eliminate `infoOffset`.

Reviewed By: avp

Differential Revision: D54656982

fbshipit-source-id: 202a1a7a2fe766d3e3406061522ef7d12a64d5f3