Wait for tests scripts to start
Summary:
The CDP Agent uses two queues for runtime tasks: one managed by
the integrator, and one inside the Async Debug API.

In tests, `scheduleScript` adds script execution to only the integrator
task queue. If a `Runtime.enable` is sent soon after, it's added to both
queues: after the scheduled script in the integrator's queue, but at
the front of the Async Debug API queue. If the async debug API queue
is wakes up first, the `Runtime.enable` will execute before the
scheduled script has run, and the `scriptParsed` event expected by
the test won't be produced.

This isn't an actual bug today, as the structure of the tests currently
guarantees that the the Async Debug API won't get a chance to run
first. The next diff in the stack changes this behavior, so this diff is
required if the next one lands. Even if the next one doesn't land,
I'd be happy to land this one, as it more closely represents real-
world use, where we can't rely on which task queue will execute
a particular task.

Reviewed By: dannysu

Differential Revision: D54808868

fbshipit-source-id: 376775dfc4596735aeba1b558b2334ab5a2d0726