Populate scopes on calls in sibling functions
Summary:
Perform propagation of the scope operand to a call when the closure
being called is known to be loaded from its parent scope.

This allows us to avoid GetClosureScope when inlining into sibling
functions that obtain the closure from the frame, since they can just
reuse the scope operand to the LoadFrameInst.

Reviewed By: avp

Differential Revision: D54837411

fbshipit-source-id: 0b2acdd13be83c6abef0dcb39e10f7a6d5766eb6