Encode environment size in CreateFunctionEnvironment
Summary:
Record the environment size as an operand to scope creation in HBC.
This allows us to decouple information about the scope from the
function. It also unlocks potential size wins, since the scope
instruction can be optimised out (unlike the function header), and we
can potentially shrink the function header significantly (see later
diffs).

Now that the size is part of the instruction, stop dumping it in the
function header (it will be removed later in the stack).

Reviewed By: avp

Differential Revision: D54573383

fbshipit-source-id: c88e300dce694f6e822642b6daca11e6f74741e4