Remove infoOffset from function header
Summary:
Our current format for function headers involves a compact header that
can overflow turn into an offset pointing to a larger out-of-line
header. Both types of header also contain an `infoOffset` field, which
points to the location of exception handler and debug information.

However, we actually place the out-of-line header and the EH/debug
information in the same "function info" region. This means that if we
treat headers with EH/debug information as overflows, we can eliminate
the `infoOffset` field, instead using the existing overflow mechanism
to point to that information.

This means that for functions with EH/debug information, we will always
also have to emit the out-of-line function header, but given that these
are not common in a release build, this is a significant size reduction
in the header. Coupled with the previous change to eliminate the
environment size, the header shrinks from 16 bytes to 12 bytes. (note
that the header size needs to be a multiple of 4 bytes, so both changes
are necessary to get any win)

Reviewed By: avp

Differential Revision: D54573385

fbshipit-source-id: c5f555d9aa9b234afb387d7cfb9acfe90613e6f1